# src/pi_control_hub_api/models/device_command.py
# coding: utf-8

"""
    PiControl Hub API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictBytes, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Optional, Union
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class DeviceCommand(BaseModel):
    """
    DeviceCommand
    """ # noqa: E501
    pairing_id: Optional[StrictStr] = Field(default=None, description="The pairing ID of the paired device.", alias="pairingId")
    driver_id: Optional[StrictStr] = Field(default=None, description="The driver ID of the paired device.", alias="driverId")
    device_id: Optional[StrictStr] = Field(default=None, description="The ID of the paired device.", alias="deviceId")
    command_id: Optional[StrictInt] = Field(default=None, description="The command ID.", alias="commandId")
    name: Optional[StrictStr] = Field(default=None, description="The command name.")
    icon: Optional[Union[StrictBytes, StrictStr]] = Field(default=None, description="Base64 encoded data of the command icon.")
    __properties: ClassVar[List[str]] = ["pairingId", "driverId", "deviceId", "commandId", "name", "icon"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of DeviceCommand from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of DeviceCommand from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "pairingId": obj.get("pairingId"),
            "driverId": obj.get("driverId"),
            "deviceId": obj.get("deviceId"),
            "commandId": obj.get("commandId"),
            "name": obj.get("name"),
            "icon": obj.get("icon")
        })
        return _obj