# src/pi_control_hub_api/models/device_driver.py
# coding: utf-8

"""
    PiControl Hub API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class DeviceDriver(BaseModel):
    """
    DeviceDriver
    """ # noqa: E501
    driver_id: Optional[StrictStr] = Field(default=None, description="The ID of the device driver.", alias="driverId")
    display_name: Optional[StrictStr] = Field(default=None, description="Name of the device driver; might be displayed in a UI.", alias="displayName")
    description: Optional[StrictStr] = Field(default=None, description="Description for the plugin that might be displayed in a UI.")
    authentication_method: Optional[StrictStr] = Field(default=None, description="The authentication method to be used when connecting to a device provided by the driver.", alias="authenticationMethod")
    __properties: ClassVar[List[str]] = ["driverId", "displayName", "description", "authenticationMethod"]

    @field_validator('authentication_method')
    def authentication_method_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value

        if value not in ('NONE', 'PIN', 'PASSWORD', 'USER_AND_PASSWORD'):
            raise ValueError("must be one of enum values ('NONE', 'PIN', 'PASSWORD', 'USER_AND_PASSWORD')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of DeviceDriver from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of DeviceDriver from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "driverId": obj.get("driverId"),
            "displayName": obj.get("displayName"),
            "description": obj.get("description"),
            "authenticationMethod": obj.get("authenticationMethod")
        })
        return _obj